Remove outdated HomeOld and HomeOld2 components to streamline codebase and improve maintainability.
